fix: modified nextauth route changed nextauth file structure to comply with next build app router rules
