+++ src/app/api/auth/[...nextauth]/getAuth.ts
@@ -1,0 +1,6 @@
+import { getServerSession } from "next-auth";
+import { authOptions } from "./authOptions";
+
+export const getAuth = () => {
+  return getServerSession(authOptions);
+};
