+++ src/app/dashboard/page.tsx
@@ -1,6 +1,6 @@
 import { redirect } from "next/navigation";
 import { StickyNotes } from "./StickyNote";
-import { getAuth } from "../api/auth/[...nextauth]/route";
+import { getAuth } from "../api/auth/[...nextauth]/getAuth";
 import { trpc } from "../api/trpc/[trpc]/trpcClient";
 
 export default async function Dashboard() {
